Redesign ProductListPage
